fix: redux 관련 코드  제거
