+++ src/app/profile/appliedStudies/AppliedStudiesContainer.tsx
@@ -1,13 +1,10 @@
-import { useAppDispatch, useAppSelector } from '@src/hooks/appRedux';
-import { setAppliedStudies } from '@src/store/modules/user';
+
 import * as Factory from 'factory.ts';
 import React, { useEffect } from 'react';
 import { AppliedStudyInterface } from '../interface/interface';
 import AppliedStudiesPresenter from './AppliedStudiesPresenter';
 
 const AppliedStudiesContainer = (): JSX.Element => {
-	const dispatch = useAppDispatch();
-	const appliedStudies = useAppSelector((state) => state.user.appliedStudies);
 
 	const getAppliedStudiesData = (iter: number) => {
 		const appliedStudiesFactory = Factory.Sync.makeFactory<AppliedStudyInterface>({
@@ -23,20 +20,11 @@ const AppliedStudiesContainer = (): JSX.Element => {
 		return appliedStudiesFactory.buildList(iter);
 	};
 
-	useEffect(() => {
-		dispatch(
-			setAppliedStudies({
-				openedAppliedStudies: getAppliedStudiesData(5),
-				closedAppliedStudies: getAppliedStudiesData(5),
-			})
-		);
-	}, []);
-
 	return (
 		<div>
 			<AppliedStudiesPresenter
-				openedAppliedStudies={appliedStudies?.openedAppliedStudies}
-				closedAppliedStudies={appliedStudies?.closedAppliedStudies}
+				openedAppliedStudies={getAppliedStudiesData(5)}
+				closedAppliedStudies={getAppliedStudiesData(3)}
 			/>
 		</div>
 	);
