+++ src/app/profile/bookmark/BookmarkContainer.tsx
@@ -1,13 +1,9 @@
 import React, { useEffect } from 'react';
 import * as Factory from 'factory.ts';
-import { setBookmarks } from '@src/store/modules/user';
-import { useAppDispatch, useAppSelector } from '@src/hooks/appRedux';
 import BookmarkPresenter from './BookmarkPresenter';
 import { BookmarkInterface } from '../interface/interface';
 
 const BookmarkContainer = (): JSX.Element => {
-	const dispatch = useAppDispatch();
-	const bookmarks = useAppSelector((state) => state.user.bookmarks);
 
 	const getBookmarkData = (iter: number) => {
 		const bookmarkFactory = Factory.Sync.makeFactory<BookmarkInterface>({
@@ -23,20 +19,11 @@ const BookmarkContainer = (): JSX.Element => {
 		return bookmarkFactory.buildList(iter);
 	};
 
-	useEffect(() => {
-		dispatch(
-			setBookmarks({
-				recruitingBookmarks: getBookmarkData(5),
-				recruitedBookmarks: getBookmarkData(3),
-			})
-		);
-	}, []);
-
 	return (
 		<div>
 			<BookmarkPresenter
-				recruitedBookmarks={bookmarks?.recruitedBookmarks}
-				recruitingBookmarks={bookmarks?.recruitingBookmarks}
+				recruitedBookmarks={getBookmarkData(5)}
+				recruitingBookmarks={getBookmarkData(3)}
 			/>
 		</div>
 	);
