+++ src/app/profile/my/notice/NoticeDetailPresenter.tsx
@@ -1,8 +1,8 @@
-import { NoticeInterface } from '@src/store/modules/user';
 import moment from 'moment';
 import React from 'react';
 import { IoArrowBack } from 'react-icons/io5';
 import { useHistory } from 'react-router-dom';
+import { NoticeInterface } from '../../interface/interface';
 import './styles.scss';
 
 interface NoticeDetailPresenterProps {
