+++ src/app/profile/recruitingStudies/RecruitingStudiesContainer.tsx
@@ -1,14 +1,9 @@
-import { useAppDispatch, useAppSelector } from '@src/hooks/appRedux';
 import React, { useEffect } from 'react';
 import * as Factory from 'factory.ts';
-import { setRecruitingStudies } from '@src/store/modules/user';
 import RecruitingStudiesPresenter from './RecruitingStudiesPresenter';
 import { AppliedStudyInterface } from '../interface/interface';
 
 const RecruitingStudiesContainer = (): JSX.Element => {
-	const dispatch = useAppDispatch();
-	const recruitingStudies = useAppSelector((state) => state.user.recruitingStudies);
-
 	const getRecruitingStudiesData = (iter: number, status: string) => {
 		const recruitingStudiesFactory = Factory.Sync.makeFactory<AppliedStudyInterface>({
 			studyId: Factory.each((i) => i),
@@ -23,20 +18,11 @@ const RecruitingStudiesContainer = (): JSX.Element => {
 		return recruitingStudiesFactory.buildList(iter);
 	};
 
-	useEffect(() => {
-		dispatch(
-			setRecruitingStudies({
-				openedRecruitingStudies: getRecruitingStudiesData(5, 'recruiting'),
-				closedRecruitingStudies: getRecruitingStudiesData(3, 'closed'),
-			})
-		);
-	}, []);
-
 	return (
 		<div>
 			<RecruitingStudiesPresenter
-				openedRecruitingStudies={recruitingStudies?.openedRecruitingStudies}
-				closedRecruitingStudies={recruitingStudies?.closedRecruitingStudies}
+				openedRecruitingStudies={getRecruitingStudiesData(5, 'recruiting')}
+				closedRecruitingStudies={getRecruitingStudiesData(3, 'closed')}
 			/>
 		</div>
 	);
