+++ src/app/shared/components/input/InputBase.tsx
@@ -1,6 +1,7 @@
-import { useAppSelector } from '@src/hooks/appRedux';
 import React from 'react';
 import './inputBase.scss';
+
+const sampleImgUrl = 'https://cdn.pixabay.com/photo/2016/05/05/02/37/sunset-1373171__480.jpg';
 
 interface InputBaseProps {
 	placeholder: string;
@@ -8,7 +9,6 @@ interface InputBaseProps {
 	setContent: (content: string) => void;
 }
 const InputBase = ({ placeholder, content, setContent }: InputBaseProps): JSX.Element => {
-	const profilePicture = useAppSelector((state) => state.user.userInfo.profilePicture);
 
 	const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
 		setContent(event.target.value);
@@ -16,7 +16,7 @@ const InputBase = ({ placeholder, content, setContent }: InputBaseProps): JSX.El
 
 	return (
 		<div className="inputbase-container">
-			<img className="inputbase-img" src={profilePicture} alt="프로필 이미지" width={40} height={40} />
+			<img className="inputbase-img" src={sampleImgUrl} alt="프로필 이미지" width={40} height={40} />
 			<input className="inputbase-input" placeholder={placeholder} value={content} onChange={onChange} />
 		</div>
 	);
