--- src/hooks/appRedux.ts
@@ -1,5 +1,0 @@
-import { AppDispatch, RootState } from '@src/store';
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-
-export const useAppDispatch = () => useDispatch<AppDispatch>();
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
