+++ src/index.tsx
@@ -1,16 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from '@src/App';
-import { Provider } from 'react-redux';
-import { store } from './store';
 import reportWebVitals from './reportWebVitals';
 import './index.scss';
 
 ReactDOM.render(
 	<React.StrictMode>
-		<Provider store={store}>
-			<App />
-		</Provider>
+		<App />
 	</React.StrictMode>,
 	document.getElementById('root')
 );
