--- src/store/index.ts
@@ -1,13 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import logger from 'redux-logger';
-import userReducer from './modules/user';
-
-export const store = configureStore({
-	reducer: {
-		user: userReducer,
-	},
-	middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
-});
-
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
